requirements.txt
accelerate==0.9.0
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
denoising-diffusion-pytorch==1.9.4
docker-pycreds==0.4.0
einops==0.4.1
ema-pytorch==0.2.4
gitdb==4.0.9
GitPython==3.1.18
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pathlib==1.0.1
pathtools==0.1.2
Pillow==8.4.0
protobuf==3.19.6
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
PyYAML==6.0.1
requests==2.27.1
scipy==1.5.4
sentry-sdk==1.32.0
setproctitle==1.2.3
six==1.16.0
smmap==5.0.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.18
wandb==0.15.12
zipp==3.6.0
